SOF-677: Mer robust beregning mtp veldig store (positive og negative) beløp

+++ uttrykk-kjerne/src/main/java/ske/fastsetting/skatt/uttrykk/belop/KroneUttrykk.java
@@ -30,7 +30,7 @@ public class KroneUttrykk extends AbstractUttrykk<Belop, KroneUttrykk> implement
     }
 
     public static KroneUttrykk kr(BigInteger belop) {
-        return new KroneUttrykk(Belop.kr(belop.intValue()),Type.MedOre);
+        return new KroneUttrykk(Belop.kr(belop),Type.MedOre);
     }
 
     public static KroneUttrykk kr(double belop) {
